requirements.txt
# HTTP and networking
requests>=2.28.0
PySocks>=1.7.1
stem>=1.8.0

# HTML parsing
beautifulsoup4>=4.11.0
lxml>=4.9.0

# Image processing
Pillow>=9.0.0

# Machine learning models
torch>=2.0.0
transformers>=4.25.0
accelerate>=0.20.0  # For faster model loading

# Text processing
llmlingua>=0.1.0  # Optional: for text simplification

# Additional dependencies
tqdm>=4.66.0  # For progress bars